refactor: 파일 위치를 directory로 묶어서 새 경로로 위치
