+++ src/main/java/com/seniorway/seniorway/config/jwt/JwtAuthenticationFilter.java
@@ -1,12 +1,10 @@
-package com.seniorway.seniorway.config;
+package com.seniorway.seniorway.config.jwt;
 
-import com.seniorway.seniorway.config.JwtTokenProvider;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 import org.springframework.web.filter.OncePerRequestFilter;
 
 import java.io.IOException;
