+++ src/main/java/com/seniorway/seniorway/config/jwt/SecurityConfig.java
@@ -1,6 +1,5 @@
-package com.seniorway.seniorway.config;
+package com.seniorway.seniorway.config.jwt;
 
-import com.seniorway.seniorway.config.JwtTokenProvider;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
