+++ src/main/java/com/seniorway/seniorway/dto/user/UserSignUpRequestsDto.java
@@ -1,4 +1,4 @@
-package com.seniorway.seniorway.dto;
+package com.seniorway.seniorway.dto.user;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
